+++ OUDS/Core/Components/Sources/Controls/Checkbox/OUDSCheckbox.swift
@@ -94,7 +94,7 @@ public struct OUDSCheckbox: View {
 
     // MARK: Properties
 
-    private let a11yLabel: String
+    private let accessibilityLabel: String
     private let isError: Bool
     private let isReadOnly: Bool
 
@@ -123,7 +123,7 @@ public struct OUDSCheckbox: View {
             OL.warning("The OUDSCheckbox should not have an empty accessibility label, think about your disabled users!")
         }
         _isOn = isOn
-        a11yLabel = accessibilityLabel
+        self.accessibilityLabel = accessibilityLabel
         self.isError = isError
         self.isReadOnly = isReadOnly
     }
@@ -143,9 +143,9 @@ public struct OUDSCheckbox: View {
                 .modifier(CheckboxBackgroundColorModifier(interactionState: interactionState))
         }
         .accessibilityRemoveTraits([.isButton]) // .isToggle trait for iOS 17+
-        .accessibilityLabel(a11yLabel(isDisabled: !isEnabled))
-        .accessibilityValue(a11yValue())
-        .accessibilityHint(a11yHint())
+        .accessibilityLabel(accessibilityLabel)
+        .accessibilityValue(accessibilityValue)
+        .accessibilityHint(accessibilityHint)
     }
 
     // MARK: Computed value
@@ -156,13 +156,18 @@ public struct OUDSCheckbox: View {
 
     // MARK: - A11Y helpers
 
-    /// Forges a string to vocalize with *Voice Over* describing the component value
-    private func a11yValue() -> String {
-        isOn ? "core_checkbox_checked_a11y".localized() : "core_checkbox_unchecked_a11y".localized()
+    /// Forges a string to vocalize with *Voice Over* describing the component trait, value, state and error
+    private var accessibilityValue: String {
+        let traitDescription = "core_checkbox_trait_a11y".localized() // Fake trait for Voice Over vocalization
+        let valueDescription = isOn ? "core_checkbox_checked_a11y".localized() : "core_checkbox_unchecked_a11y".localized()
+        let stateDescription = !isEnabled || isReadOnly ? "core_common_disabled_a11y".localized() : ""
+        let errorDescription = isError ? "core_common_onError_a11y".localized() : ""
+
+        return "\(traitDescription). \(valueDescription). \(stateDescription). \(errorDescription)"
     }
 
     /// Forges a string to vocalize with *Voice Over* describing the component hint
-    private func a11yHint() -> String {
+    private var accessibilityHint: String {
         if !isEnabled || isReadOnly {
             ""
         } else {
@@ -170,16 +175,5 @@ public struct OUDSCheckbox: View {
                 ? "core_checkbox_hint_a11y" <- "core_checkbox_unchecked_a11y".localized()
                 : "core_checkbox_hint_a11y" <- "core_checkbox_checked_a11y".localized()
         }
-    }
-
-    /// Forges a string to vocalize with *Voice Over* describing the component state
-    /// - Parameter isDisabled: True if component is disabled, false otherwise
-    private func a11yLabel(isDisabled: Bool) -> String {
-        let stateDescription = isDisabled || isReadOnly ? "core_common_disabled_a11y".localized() : ""
-        let errorDescription = isError ? "core_common_onError_a11y".localized() : ""
-        let checkboxA11yTrait = "core_checkbox_trait_a11y".localized() // Fake trait for Voice Over vocalization
-
-        let result = "\(a11yLabel), \(stateDescription), \(errorDescription), \(checkboxA11yTrait)"
-        return result
     }
 }