+++ OUDS/Core/Components/Sources/Controls/Checkbox/OUDSCheckboxItem.swift
@@ -241,9 +241,9 @@ public struct OUDSCheckboxItem: View {
     public var body: some View {
         ControlItem(indicatorType: .checkBox(convertedState), layoutData: layoutData, action: action)
             .accessibilityRemoveTraits([.isButton]) // .isToggle trait for iOS 17+
-            .accessibilityLabel(a11yLabel)
-            .accessibilityValue(a11yValue)
-            .accessibilityHint(a11yHint)
+            .accessibilityLabel(accessibilityLabel)
+            .accessibilityValue(accessibilityValue)
+            .accessibilityHint(accessibilityHint)
     }
 
     // MARK: - Computed value
@@ -254,30 +254,28 @@ public struct OUDSCheckboxItem: View {
 
     // MARK: - A11Y helpers
 
-    /// The text to vocalize with *Voice Over* for the state of the indicator
-    private var a11yValue: String {
-        (isOn ? "core_checkbox_checked_a11y" : "core_checkbox_unchecked_a11y").localized()
+    /// Forge a string to vocalize the component label based on label, extraLabel and description
+    private var accessibilityLabel: String {
+        let extraLabel = layoutData.extraLabel?.isEmpty != false ? "" : ", \(layoutData.extraLabel ?? "")"
+        let description = layoutData.description?.isEmpty != false ? "" : ", \(layoutData.description ?? "")"
+        return "\(layoutData.label)\(extraLabel)\(description)"
     }
 
-    /// Forges a string to vocalize with *Voice Over* describing the component state.
-    private var a11yLabel: String {
-        let stateDescription: String = layoutData.isReadOnly || !isEnabled ? "core_common_disabled_a11y".localized() : ""
+    /// Forges a string to vocalize with *Voice Over* describing the component trait, value, state and error
+    private var accessibilityValue: String {
+        let traitDescription = "core_checkbox_trait_a11y".localized() // Fake trait for Voice Over vocalization
+        let valueDescription = isOn ? "core_checkbox_checked_a11y".localized() : "core_checkbox_unchecked_a11y".localized()
+        let stateDescription = !isEnabled || layoutData.isReadOnly ? "core_common_disabled_a11y".localized() : ""
 
-        var errorDescription = ""
-        if layoutData.isError {
-            let errorPrefix = "core_common_onError_a11y".localized()
-            let errorText = layoutData.errorText?.localized() ?? ""
-            errorDescription = "\(errorPrefix), \(errorText)"
-        }
+        let errorPrefix = "core_common_onError_a11y".localized()
+        let errorText = layoutData.errorText?.localized() ?? ""
+        let errorDescription = layoutData.isError ? "\(errorPrefix), \(errorText)" : ""
 
-        let checkboxA11yTrait = "core_checkbox_trait_a11y".localized() // Fake trait for Voice Over vocalization
-
-        let result = "\(stateDescription), \(layoutData.label), \(layoutData.description ?? ""), \(errorDescription), \(checkboxA11yTrait)"
-        return result
+        return "\(traitDescription). \(valueDescription). \(stateDescription). \(errorDescription)"
     }
 
-    /// Forges a string to vocalize with *Voice Over* explaining the hint for the user about the component.
-    private var a11yHint: String {
+    /// Forges a string to vocalize with *Voice Over* describing the component hint
+    private var accessibilityHint: String {
         if layoutData.isReadOnly || !isEnabled {
             ""
         } else {